# contracts/openapi/pramaan-backend.yaml
openapi: 3.1.0
info:
  title: Pramaan Next — Backend (Phase 1)
  version: 0.1.0
  description: >
    Evidence ingestion, versioning, integrity, deterministic repository queries,
    application metadata, and the collection scheduler. Phase 1 uses deterministic
    mock enterprise integrations. No LLM. Source of truth for backend ⇄ frontend
    and backend ⇄ agents interfaces — implement against this file, do not hand-roll.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/controls/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [controls]
      summary: Get one control
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ControlView" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/evidence/ingest:
    post:
      tags: [evidence]
      summary: Ingest a single evidence item (SHA-256 dedup, immutable versioning)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/IngestRequest" } } }
      responses:
        "201": { description: Ingested, content: { application/json: { schema: { $ref: "#/components/schemas/IngestResult" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/evidence/bulk:
    post:
      tags: [evidence]
      summary: Bulk ingest (JSON array). Per-item outcomes; partial success allowed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/IngestRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/BulkIngestResponse" } } } }
  /api/v1/evidence/bulk/upload:
    post:
      tags: [evidence]
      summary: Bulk upload files — each file becomes one evidence item
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [applicationSlug, framework, controlId, files]
              properties:
                applicationSlug: { type: string }
                framework: { type: string }
                controlId: { type: string }
                sourceSystem: { type: string, default: BULK_UPLOAD }
                collectedBy: { type: string }
                technology: { type: string, description: "optional; sets the canonical technology tag. Falls back to source-system inference when omitted." }
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/BulkIngestResponse" } } } }
...
  /api/v1/evidence:
    get:
      tags: [evidence]
      summary: Deterministic filtered evidence query (stable ordering)
      parameters:
        - { name: applicationSlug, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
        - { name: controlId, in: query, schema: { type: string } }
        - { name: sourceSystem, in: query, schema: { type: string } }
        - { name: collectedAfter, in: query, schema: { type: string, format: date-time } }
        - { name: collectedBefore, in: query, schema: { type: string, format: date-time } }
        - { name: tag, in: query, description: "key or key:value", schema: { type: string } }
        - { name: technology, in: query, description: "canonical technology tag facet (e.g. postgresql)", schema: { type: string } }
        - { name: collectionMethod, in: query, description: "canonical collectionMethod tag facet (scheduled | manual | bulk | on-demand)", schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 500 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/EvidenceView" }
  /api/v1/evidence/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [evidence]
      summary: Get one evidence record with its current version
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceView" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/evidence/{id}/lifecycle:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [evidence]
      summary: "Use Case 13 — evidence lifecycle state, retention/expiry and the transition audit trail"
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceLifecycleView" } } } }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [evidence]
      summary: "Transition evidence lifecycle state (SUBMIT | APPROVE | REJECT | RETIRE | RESET)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [SUBMIT, APPROVE, REJECT, RETIRE, RESET] }
                actor: { type: string }
                note: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceLifecycleView" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/evidence/{id}/versions:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [evidence]
      summary: List all versions of an evidence record
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EvidenceVersionView" } }
  /api/v1/evidence/{id}/versions/{version}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: version, in: path, required: true, schema: { type: integer } }
    get:
      tags: [evidence]
      summary: Get one version
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceVersionView" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/evidence/{id}/verify:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: version, in: query, required: false, schema: { type: integer } }
    get:
      tags: [evidence]
      summary: Recompute SHA-256 from the object store and compare
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/IntegrityReport" } } } }
  /api/v1/evidence/query/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
          enum: [source-breakdown, stale-evidence, latest-per-control, duplicates, freshness]
      - { name: applicationSlug, in: query, schema: { type: string } }
      - { name: framework, in: query, schema: { type: string } }
      - { name: controlId, in: query, schema: { type: string } }
      - { name: sourceSystem, in: query, schema: { type: string } }
    get:
      tags: [evidence]
      summary: Named deterministic query — structured answer plus rows and counts
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DeterministicQueryResult" } } } }
        "400": { $ref: "#/components/responses/Error" }
...
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    IngestRequest:
      type: object
      required: [applicationSlug, controlId, framework, sourceSystem]
      description: Exactly one of contentBase64 / contentText is required.
      properties:
        applicationSlug: { type: string }
        controlId: { type: string }
        framework: { type: string }
        sourceSystem: { type: string }
        sourceObjectId: { type: string }
        title: { type: string }
        contentType: { type: string }
        contentBase64: { type: string }
        contentText: { type: string }
        collectedAt: { type: string, format: date-time }
        collectedBy: { type: string }
        metadata: { type: object, additionalProperties: { type: string } }
        tags: { type: object, additionalProperties: { type: string } }
    IngestResult:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        evidenceKey: { type: string }
        applicationSlug: { type: string }
        controlId: { type: string }
        framework: { type: string }
        sourceSystem: { type: string }
        outcome: { type: string, enum: [CREATED, NEW_VERSION, DUPLICATE] }
        version: { type: integer }
        sha256: { type: string }
        sizeBytes: { type: integer, format: int64 }
    BulkIngestResponse:
      type: object
      properties:
        received: { type: integer }
        created: { type: integer }
        newVersions: { type: integer }
        duplicates: { type: integer }
        failed: { type: integer }
        results: { type: array, items: { $ref: "#/components/schemas/IngestResult" } }
        errors:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { type: string }
    EvidenceVersionView:
      type: object
      properties:
        version: { type: integer }
        sha256: { type: string }
        contentType: { type: string }
        sizeBytes: { type: integer, format: int64 }
        objectKey: { type: string }
        collectedAt: { type: string, format: date-time }
        collectedBy: { type: string }
        ingestionRunId: { type: string, format: uuid, nullable: true }
        metadata: { type: object, additionalProperties: { type: string } }
    EvidenceView:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        evidenceKey: { type: string }
        applicationSlug: { type: string }
        controlId: { type: string }
        framework: { type: string }
        sourceSystem: { type: string }
        sourceObjectId: { type: string, nullable: true }
        title: { type: string, nullable: true }
        currentVersion: { type: integer }
        lifecycleState: { type: string, enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, EXPIRED, SUPERSEDED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        tags: { type: object, additionalProperties: { type: string } }
        latest: { $ref: "#/components/schemas/EvidenceVersionView" }
    EvidenceLifecycleView:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        applicationSlug: { type: string }
        controlId: { type: string }
        state: { type: string, enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, EXPIRED, SUPERSEDED] }
        effectiveState: { type: string, description: "state, except APPROVED past retention reads as EXPIRED" }
        expired: { type: boolean }
        reviewedBy: { type: string, nullable: true }
        reviewedAt: { type: string, format: date-time, nullable: true }
        note: { type: string, nullable: true }
        retentionDays: { type: integer }
        ageDays: { type: integer, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        currentVersion: { type: integer }
        history:
          type: array
          items:
            type: object
            properties:
              fromState: { type: string, nullable: true }
              toState: { type: string }
              action: { type: string }
              actor: { type: string, nullable: true }
              note: { type: string, nullable: true }
              occurredAt: { type: string, format: date-time }
...
    IntegrityReport:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        version: { type: integer }
        expectedSha256: { type: string }
        actualSha256: { type: string, nullable: true }
        intact: { type: boolean }
        detail: { type: string }
...
    DeterministicQueryResult:
      type: object
      properties:
        name: { type: string }
        answerText: { type: string }
        generatedAt: { type: string, format: date-time }
        counts: { type: object, additionalProperties: true }
        rows:
          type: array
          items: { type: object, additionalProperties: true }
...
    ControlView:
      type: object
      properties:
        id: { type: string }
        framework: { type: string }
        title: { type: string }
        domain: { type: string }
...